refactor: centralize user types
