+++ learning-games/src/components/AnalyticsTracker.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useContext } from 'react'
 import { useLocation } from 'react-router-dom'
 import { UserContext } from '../shared/UserContext'
-import type { UserContextType } from '../shared/types/user'
+import type { UserContextType } from '../../../shared/types/user'
 import { getApiBase } from '../utils/api'
 
 function getCookie(name: string) {
