+++ learning-games/src/components/layout/ProgressSidebar.tsx
@@ -2,7 +2,7 @@ import { useContext, useEffect, useRef, useState } from 'react'
 import confetti from 'canvas-confetti'
 import { Link, useLocation } from 'react-router-dom'
 import { UserContext } from '../../shared/UserContext'
-import type { UserContextType } from '../../shared/types/user'
+import type { UserContextType } from '../../../../shared/types/user'
 import Tooltip from '../ui/Tooltip'
 import { getTotalPoints } from '../../utils/user'
 import { GOAL_POINTS } from '../../constants/progress'
