+++ learning-games/src/pages/LeaderboardPage.tsx
@@ -2,7 +2,7 @@ import { useContext, useMemo, useState } from 'react'
 import { Link } from 'react-router-dom'
 import { notify } from '../shared/notify'
 import { UserContext } from '../shared/UserContext'
-import type { UserContextType } from '../shared/types/user'
+import type { UserContextType } from '../../../shared/types/user'
 import { useLeaderboards } from '../shared/useLeaderboards'
 import ProgressSidebar from '../components/layout/ProgressSidebar'
 import './LeaderboardPage.css'
