+++ learning-games/src/pages/QuizGame.tsx
@@ -6,7 +6,7 @@ import { notify } from '../shared/notify'
 import { Link, useNavigate } from 'react-router-dom'
 import CompletionModal from '../components/ui/CompletionModal'
 import { UserContext } from '../shared/UserContext'
-import type { UserContextType } from '../shared/types/user'
+import type { UserContextType } from '../../../shared/types/user'
 import './QuizGame.css'
 import InstructionBanner from '../components/ui/InstructionBanner'
 import { HALLUCINATION_EXAMPLES } from '../data/hallucinationExamples'
