+++ learning-games/src/shared/UserContext.ts
@@ -1,5 +1,5 @@
 import { createContext } from 'react'
-import type { UserData, UserContextType } from './types/user'
+import type { UserData, UserContextType } from '../../../shared/types/user'
 
 export const defaultUser: UserData = {
   id: '',
